🎨 feat:Error 처리
